+++ index.html
@@ -228,7 +228,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src = "js/main.js"></script>
+  <script src = "js/main.js" type = "module"></script>
 
 </body>
 </html>
